wod and payload working

+++ obc/include/Payload.h
@@ -11,15 +11,15 @@
 class Payload {
    public:
     struct payload_data {
-        uint16_t R;
-        uint16_t S;
-        uint16_t T;
-        uint16_t U;
-        uint16_t V;
-        uint16_t W;
-        int16_t lat;
-        int16_t lon;
-        int16_t alt;
+        float R;
+        float S;
+        float T;
+        float U;
+        float V;
+        float W;
+        double lat;
+        double lon;
+        double alt;
     };
 
     typedef struct payload_data payload_data_t;
